refactor(ui): Move create user functionality to its own page
This is a part of refactoring the layout and navigation of the
admin dashboard.
